# Speed up/reduce allocations by using strings.Map